minor_errno

+++ Include/CORBA/CORBA.h
@@ -60,13 +60,13 @@ namespace Nirvana {
 /// Vendor Minor Codeset ID.
 const uint32_t VMCID = 0xFFFFF;
 
-inline uint32_t make_minor_errno (int err)
+inline uint32_t make_minor_errno (int_fast16_t err)
 {
 	assert (err <= 0xFFF);
 	return (VMCID << 12) | err;
 }
 
-inline int get_minor_errno (uint32_t minor)
+inline int_fast16_t get_minor_errno (uint32_t minor)
 {
 	if (minor >> 12 == VMCID)
 		return minor & 0xFFF;
